handler_pbcol {progressr}	R Documentation

Progression Handler: Progress Reported as an ANSI Background Color in the Terminal

Description

Progression Handler: Progress Reported as an ANSI Background Color in the Terminal

Usage

handler_pbcol(
  adjust = 0,
  pad = 1L,
  complete = function(s) crayon::bgBlue(crayon::white(s)),
  incomplete = function(s) crayon::bgCyan(crayon::white(s)),
  intrusiveness = getOption("progressr.intrusiveness.terminal", 1),
  target = "terminal",
  ...
)

Arguments

adjust
(numeric) The adjustment of the progress update, where adjust = 0 positions the message to the very left, and adjust = 1 positions the message to the very right.

pad
(integer) Amount of padding on each side of the message, where padding is done by spaces.

complete, incomplete
(function) Functions that take "complete" and "incomplete" strings that comprise the progress bar as input and annotate them to reflect their two different parts. The default is to annotation them with two different background colors and the same foreground color using the crayon package.

intrusiveness
(numeric) A non-negative scalar on how intrusive (disruptive) the reporter to the user.

target
(character vector) Specifies where progression updates are rendered.

...
Additional arguments passed to make_progression_handler().

Requirements

This progression handler requires the crayon package.

Examples

handlers(handler_pbcol)
with_progress({ y <- slow_sum(1:10) })
print(y)

[Package progressr version 0.9.0 Index]