handler_rpushbullet {progressr}	R Documentation

Progression Handler: Progress Reported via the Pushbullet Messaging Service

Description

A progression handler for RPushbullet::pbPost() of the RPushbullet package.

Usage

handler_rpushbullet(
  intrusiveness = getOption("progressr.intrusiveness.rpushbullet", 5),
  target = "gui",
  ...,
  title = "Progress update from R",
  recipients = NULL,
  email = NULL,
  channel = NULL,
  apikey = NULL,
  devices = NULL
)

Arguments

intrusiveness
(numeric) A non-negative scalar on how intrusive (disruptive) the reporter to the user.

target
(character vector) Specifies where progression updates are rendered.

title
The title of the note being posted.

recipients
A character or numeric vector indicating the devices this post should go to. If missing, the default device is looked up from an optional setting, and if none has been set the push is sent to all devices.

email
An alternative way to specify a recipient is to specify an email address. If both recipients and email are present, recipients is used.

channel
A channel tag used to specify the name of the channel as the recipient. If either recipients or email are present, they will take precedence over channel.

apikey
The API key used to access the service. It can be supplied as an argument here, via the global option rpushbullet.key, or via the file ~/.rpushbullet.json which is read at package initialization (and, if found, also sets the global option).

devices
The device to which this post is pushed. It can be supplied as an argument here, or via the file ~/.rpushbullet.json which is read at package initialization.

...
Additional arguments passed to make_progression_handler().

Requirements

This progression handler requires the RPushbullet package, a Pushbullet account, and configuration according to the instructions of the RPushbullet package. It also requires internet access from the computer where this progress handler is registered.

Examples

pkg <- "RPushbullet"
if (requireNamespace(pkg, quietly = TRUE)) {

  handlers("rpushbullet")
  with_progress({ y <- slow_sum(1:10) })
  print(y)
  
}

[Package progressr version 0.18.0 Index]