handler_shiny {progressr}	R Documentation

Progression Handler: Progress Reported via 'shiny' Widgets (GUI) in the HTML Browser

Description

A progression handler for shiny and shiny::withProgress().

Usage

handler_shiny(
  intrusiveness = getOption("progressr.intrusiveness.gui", 1),
  target = "gui",
  inputs = list(message = NULL, detail = "message"),
  ...
)

Arguments

intrusiveness
(numeric) A non-negative scalar on how intrusive (disruptive) the reporter to the user.

target
(character vector) Specifies where progression updates are rendered.

inputs
(named list) Specifies from what sources the Shiny progress elements 'message' and 'detail' should be updated. Valid sources are "message", "sticky_message" and "non_sticky_message", where "message" is short for c("non_sticky_message", "sticky_message"). For example, inputs = list(message = "sticky-message", detail = "message") will update the Shiny 'message' component from sticky messages only, whereas the 'detail' component is updated using any message.

...
Additional arguments passed to make_progression_handler().

Details

For most Shiny application there is little need to use this Shiny handler directly. Instead, it is sufficient to use withProgressShiny().

Requirements

This progression handler requires the shiny package.

Examples

## Not run: 
handlers(handler_shiny())
with_progress(y <- slow_sum(1:100))

## End(Not run)

[Package progressr version 0.9.0 Index]