Add or Remove a Global 'progression' Handler

Add or Remove a Global 'progression' Handler

register_global_progression_handler(action = c("add", "remove", "query"))

Arguments

action

(character string) If "add", a global handler is added. If "remove", it is removed, if it exists. If "query", checks whether a handler is registered or not.

Value

Returns TRUE if a handler is registered, otherwise FALSE. If action = "query", the value is visible, otherwise invisible.

Requirements

This function requires R (>= 4.0.0) - the version in which global calling handlers where introduces.

Examples

if (getRversion() >= "4.0.0" && !is.element("pkgdown", loadedNamespaces()))
handlers(global = TRUE)

## This renders progress updates for each of the three calls slow_sum()
for (ii in 1:3) {
  xs <- seq_len(ii + 3)
  message(sprintf("%d. slow_sum()", ii))
  y <- slow_sum(xs, stdout = TRUE, message = TRUE)
  print(y)
}
#> 1. slow_sum()
#> O: Element #1
#> M: Added value 1
#> O: Element #2
#> M: Added value 2
#> O: Element #3
#> M: Added value 3
#> O: Element #4
#> M: Added value 4
#> [1] 10
#> 2. slow_sum()
#> O: Element #1
#> M: Added value 1
#> O: Element #2
#> M: Added value 2
#> O: Element #3
#> M: Added value 3
#> O: Element #4
#> M: Added value 4
#> O: Element #5
#> M: Added value 5
#> [1] 15
#> 3. slow_sum()
#> O: Element #1
#> M: Added value 1
#> O: Element #2
#> M: Added value 2
#> O: Element #3
#> M: Added value 3
#> O: Element #4
#> M: Added value 4
#> O: Element #5
#> M: Added value 5
#> O: Element #6
#> M: Added value 6
#> [1] 21

if (getRversion() >= "4.0.0" && !is.element("pkgdown", loadedNamespaces()))
handlers(global = FALSE)